% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boundary_from_track.R
\name{create_boundary_from_track}
\alias{create_boundary_from_track}
\title{Create a boundary file from migration tracks}
\usage{
create_boundary_from_track(track_points, pu_grid)
}
\arguments{
\item{track_points}{XY coordinates of migration tracks. An sf points object in a  projected coordinate system. The sf data frame may contain a column 'id' giving the identity of the individual. The sf data frame may contain a column 'date' in YYYY-MM-DD format to order the sampling locations.}

\item{pu_grid}{A planning unit grid as an sf polygon object in a projected coordinate system}
}
\value{
A planning unit boundary file as an edge list.
}
\description{
\code{create_boundary_from_track} determines the spatial overlap between one or more migration tracks and a planning unit grid. It then generates a boundary file for use in conservation planning. Adjacent planning units which have a migration track running through them have shared boundary. The resulting boundary file can be in the format of an edge list or matrix.
}
\section{Details}{

'edge_list' has columns \emph{id1}, \emph{id2}, \emph{boundary}, and \emph{boundarylength} \cr
\itemize{
\item \emph{id1} is the ID of the first planning unit
\item \emph{id2} is the ID of the connected planning unit
\item \emph{boundary} is 1, representing a logical TRUE connection between the pair of planning units
\item \emph{boundarylength} is the length of the shared boundary in units of the projected coordinate system
}
}

